requirements.txt
Flask==3.0.0
markdown==3.5.2
python-frontmatter==1.0.0
Pygments==2.17.2
Flask-Compress==1.14
Flask-Caching==2.1.0
flask-cors==4.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
